requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
torch==2.2.0
transformers==4.35.0
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
requests==2.31.0
numpy==1.24.3
spacy==3.7.2
python-multipart==0.0.6